pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "probeai"
version = "0.1.0"
description = "Evaluation and testing framework for RAG and AI agent systems"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "pgvector>=0.2.5",
    "psycopg2-binary>=2.9.9",
    "sentence-transformers>=2.3.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "pandas>=2.2.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "api*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
